+++ utah-design-system-website/src/typedefs.d.js
@@ -244,4 +244,27 @@
  *   @property {string} popupText
  * }
  */
+/**
+ * @typedef CheckboxExamplePropsShape {
+ * @property {string} className
+ * @property {string} errorMessage
+ * @property {string} id
+ * @property {boolean} isDisabled
+ * @property {string} label
+ * @property {boolean} value
+ * }
+ */
+
+/**
+ * @typedef RadioButtonExamplePropsShape {
+ * @property {string} className
+ * @property {string} errorMessage
+ * @property {string} id
+ * @property {boolean} isDisabled
+ * @property {boolean} isRequired
+ * @property {string} label
+ * @property {boolean} value
+ * }
+ */
+
 export default false;